sdap-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdap"
version = "0.1.0"
description = "Secure Digital Agent Protocol — Python reference implementation"
requires-python = ">=3.11"
dependencies = [
    "PyJWT>=2.8",
    "cryptography>=42.0",
    "httpx>=0.27",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["sdap"]
